Optimize queries
I used a derived column via `@Formula` to get rid of the need for lazy
fetching of `Event#participants`.
I could have made that happen with a custom projection and a native
JOIN, too, but using JPQL and the derived column seemed simpler
maintenance-wise.

Therefore, I also rewrote the other `@EntityGraph`-annotated queries
with JPQL queries.

Note: The `DISTINCT` is needed to not return a row/element in the result
 set for every n in the 1-n relation between event and participants. Cf.
  https://in.relation.to/2016/08/04/introducing-distinct-pass-through
  -query-hint/.