Enforce boundary between event entity and domain classes
This basically means mapping the `Event` entity to `LostMailEvent` or
`MatchMailEvent` as soon as possible. These objects then are safe to
handle: they are immutable (if possible), their state is not subject to
transactional semantics, and they carry only the information they need.

In the same vein I reduced method and class visibility to
package-private where I hadn't already done so.