Send mail asynchronously (again)
Initially, I moved from async mail dispatch to sync because I did not
feel the benefit outweighed the added complexity. This was with a
mechanism to handle mail dispatch outcomes, however.

Now we do not care for mail dispatch outcomes anymore since we can only
handle trivial application-side errors, anyway. If we wanted to make
sure that either 1) mail reaches its destination or 2) we recover from
dispatch errors, we'd need to invest a lot more time and effort.

So, as it's basically moot to check the result of the mail dispatch, we
can make it async again.

This has the added advantage of keeping DB transactions with mail
dispatch "side effects" much shorter.