Schedule matching and notifying to exact deadline
Now, we don't periodically check for event's whose deadline has passed
and then send notifications anymore. Rather, we tell the scheduler when
exactly to do that by providing it with the event's ID and deadline.

So, for an event with a deadline of 2023-02-12T00:00:00Z, the scheduler
will try trigger the notification flow at exactly that time.

All in all a cleaner, more deterministic implementation, making use of
Spring's event mechanism. It has one flaw, however: On redeploy/restart,
we'll lose the scheduling queue. I didn't want to persist it in a DB
because that'd come with lots of drawbacks, too, plus more complexity.
Therefore, I chose the pragmatic solution of re-enqueuing the triggers
on application start and hope, there aren't too many events in the DB.

Note: I also had a look at the tests and found they were prone to fail
non-deterministically due to lack of care when persisting/deleting
through `TestEventRepository`. I now flush deliberately to make sure the
required ops (insert, delete) actually get carried out.