Display error when user attempts to register after deadline
Previously, we didn't show anything and just redirected to the event
page. Now, the user gets a proper error page.

Handling this allowed, or even suggested, we get rid of the `@Formula`'d
`c.r.w.d.Deadline.instant` column in favor of a derived property
`getInstant` since we have no use for
`c.r.w.u.r.RegisterParticipantRepository.findByIdAndDeadlineAfter`
anymore, which was the sole reason for the derived/virtual column:
Instead of returning an empty `Optional` we now have to first retrieve
the event anyway and then check its deadline in order to be able to
react appropriately, namely with a
`c.r.w.u.r.RegistrationAttemptTooLateException`.