Base local dev on Spring Boot docker compose support
Cf. https://spring.io/blog/2023/06/21/docker-compose-support-in-spring-boot-3-1.

We could get rid of the additional app bootstrapping in the test
sources and make everything a bit more declarative. Plus, no more
dynamic ports for local dev.

It came at the cost of introducing a separate profile, but it wasn't
the best of ideas to have local dev use the test "profile"
(src/test/resources/config/application.properties), too, in the first
place.