Couple "event created" event to successful transaction
Before, we would employ a manual transaction via `TransactionTemplate`.
On tx rollback due to constraint violation, we would pass on sending the
event. This 1) didn't cover other rollback reasons and 2) there's not a
1:1 match between a `ConstraintViolationException` and the "duplicate
event" problem we wanted to handle here.

Now we explicitly try and find a duplicate event first, and roll the tx
back if that's true. In this case we also pass on sending the event by
way of the `@TransactionalEventListener` semantics.

It's a bit unfortunate that the listener has to be aware of a
transaction happening at the site where the event is getting published,
but at least we can use `fallbackExecution = true`.