Load participants for registration only if required
The only place we require them is when adding a participant after we've
ensured the registration isn't a duplicate.

Also, now we're relying on db constraints instead of Set#add semantics
when detecting duplicate registration attempts. This way, we can easily
change the collection type of Event#participants without breaking
business logic.