Make registration DTO minimal
Instead of posting essentially the whole event + the registration
attempt, we now only submit the the latter through the form.

This necessitated a small "hack" for RegisterParticipantController#post:
When the DTO is not @Valid, we need to fetch the event data again
through RegisterParticipantService.getEvent and add it again to the
model.

Notes:
- RegisterParticipantService.getEvent is a new method, not caring about
the event's deadline. The old semantics have been transferred to
RegisterParticipantService.getEventOpenForRegistration. This means that
we can register a participant as long as they initially succeeded in
opening the registration page in time. Else they might have submitted an
initial, but invalid registration attempt that cannot be remedied since
they deadline has passed in the meantime.
- Added @ModelAttribute names, as I battled with name drift and thus
broken binding.