Provide poor man's at-most-once for notifications
Fetching and deleting the event is now demarcated by a very short tx.
I.e. should, in a multi-instance deployment, other instance(s) try to
notify for the same event ID, all but one instance will run into an
exception on NotifyRepository#delete and then skip the notifications
proper.

I deliberately opted for a potentially failing tx (and early exit)
instead of a lock so we don't interfere with other clients of the
entities, like RegisterParticipantService, for example.

We need to be careful not to modify the entity after the tx, but that
was also the case before, and will only change with retrieving a
projection instead of the entity itself.